learning-labs-hq/backend-tutor/requirements.txt
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
httpx==0.25.1
requests==2.31.0
gunicorn==21.2.0
